chore(banner): move NotificationIcon to common
