+++ packages/components/src/common/NotificationIcon/NotificationIcon.stories.tsx
@@ -1,8 +1,8 @@
 import * as React from "react";
 import NotificationIcon, { variantToIconAssetsMap } from "./NotificationIcon";
-import { BannerColor } from "../Banner";
+import { BannerColor } from "../../Banner";
 import { Story } from "@storybook/react/types-6-0";
-import Text from "../Text";
+import Text from "../../Text";
 import styles from "./NotificationIcon.stories.module.css";
 
 // todo (Andrew): look into getting rid of the `as` cast. The `stylesByColor` object's keys are
@@ -10,7 +10,7 @@ import styles from "./NotificationIcon.stories.module.css";
 const variants = Object.keys(variantToIconAssetsMap) as BannerColor[];
 
 export default {
-  title: "NotificationIcon",
+  title: "common/NotificationIcon",
   component: NotificationIcon,
 };
